feat: :sparkles: Envio de mensagem no RabbitMQ
- Adição do pg admin no docker-compose
- Correção da conexão com BD

+++ banking-validation-master/src/main/java/br/com/alura/domain/Audit.java
@@ -1,18 +1,18 @@
 package br.com.alura.domain;
 
 public class Audit {
-    private Long id;
+    private Integer id;
     private String cnpj;
     private String situacaoCadastral;
 
 
-    public Audit(Long id, String cnpj, String situacaoCadastral) {
+    public Audit(Integer id, String cnpj, String situacaoCadastral) {
         this.id = id;
         this.cnpj = cnpj;
         this.situacaoCadastral = situacaoCadastral;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
